refactor: move `PictureOrientation` enum to the new `models.picture` package
