+++ src/main/kotlin/domain/models/picture/PictureOrientation.kt
@@ -1,4 +1,4 @@
-package domain.models
+package domain.models.picture
 
 /**
  * Enum representing the possible orientations for the pictures in a collage.
