+++ src/test/kotlin/domain/models/CollageConfigurationTest.kt
@@ -2,6 +2,7 @@ package domain.models
 
 import domain.exceptions.collageConfiguration.CollageConfigurationNegativeRowsOrColumnsException
 import domain.exceptions.collageConfiguration.CollageConfigurationZeroRowsOrColumnsException
+import domain.models.picture.PictureOrientation
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
